Finished GuessGame
Rather than put validation rules in the ConsoleMsg I chose to complete the count of tries and loops in the Game Events. I would have liked to include more validations such as numeral validation for the guesses but I had trouble with executing the program. I could have definitely added validations to the GameEvents Class but i did not want the program to become spaghetti code. I tried to keep the game as readable as possible. After completing this game I really feel as if my understanding of C# has improved.